networking_agent/Agents/requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
langgraph==0.0.20
openai==1.3.0
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.24.3 